deal with unicode characters, and test
fixes #6

Add some magic that tries to make everything utf-8 consistently on Python 2 and 3. Should be able to deal with unicode characters in any freeform field of manager. Most of the problems are with Python 2 so code is mostly to make Py2 more compatible.

Tests are set to run through py.test (with scripttest) and tox ensuring that we work in UTF-8 environmets throughout. Tests will break in an 'ascii' terminal since they are trying to produce unicode characters.